+++ src/main/java/kitchenpos/order/application/OrderStatusService.java
@@ -1,0 +1,40 @@
+package kitchenpos.order.application;
+
+import kitchenpos.order.dao.OrderLineItemDao;
+import kitchenpos.order.domain.Order;
+import kitchenpos.order.domain.OrderDao;
+import kitchenpos.order.domain.OrderStatus;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.Objects;
+
+@Service
+public class OrderStatusService {
+    private final OrderDao orderDao;
+    private final OrderLineItemDao orderLineItemDao;
+
+    public OrderStatusService(final OrderDao orderDao, final OrderLineItemDao orderLineItemDao) {
+        this.orderDao = orderDao;
+        this.orderLineItemDao = orderLineItemDao;
+    }
+
+    @Transactional
+    public Order changeOrderStatus(final Long orderId, final Order order) {
+        final Order savedOrder = orderDao.findById(orderId)
+                .orElseThrow(IllegalArgumentException::new);
+
+        if (Objects.equals(OrderStatus.COMPLETION.name(), savedOrder.getOrderStatus())) {
+            throw new IllegalArgumentException();
+        }
+
+        final OrderStatus orderStatus = OrderStatus.valueOf(order.getOrderStatus());
+        savedOrder.setOrderStatus(orderStatus.name());
+
+        orderDao.save(savedOrder);
+
+        savedOrder.setOrderLineItems(orderLineItemDao.findAllByOrderId(orderId));
+
+        return savedOrder;
+    }
+}
